product: reorder members to address performance regression on Moore
This makes no sense at all to me, yet it is reproducible.  The commit
2228f9dbfb2c38191d148ac6b76b09d83e0cbd95 introduces a severe
performance regression on Moore as measured by score: from 2.67s to
4.32s on erebus, Clang 3.5 -O3 -DNDEBUG.

It makes no sense, as this commit has nothing to do with Moore at all.
Even if I strip 'vcsn score' from all the tests that are before that
one (in case something would ruin the memory or whatever) I can still
see that problem.  I've repeated the measurement many times, and it is
_that_ commit, no doubt about it.

I have no idea what's going on, but another Mac I see no performance
regression at all.  A closer look at this commit shows it does two
things: one is really addressing what it claims, removing a
"weightset" variable, and another is reorganizing members to put them
at the end, as it usual.  It turns out that one another member
variable was missed (origins_), and therefore the order of the members
was changed.

This patch moves origins_ as last member of the product_automaton, and
magically enough, restores the performances.

1: v2b.3-672-g4be56f5
2: v2b.3-673-g2228f9d     (the culprit)
3: v2b.3-675, this commit

  1      2      3

4.14s  4.13s  4.09s  a & 12               # a = std([a-e]*b(<2>[a-e])*)
7.83s  7.72s  7.77s  a.determinize()      # a = ladybird(21)
0.54s  0.53s  0.53s  a.eval("a"*151)      # a = de_bruijn(150), 1000x
0.35s  0.37s  0.34s  a.format("dot")      # a = std(a?{500})
0.12s  0.13s  0.12s  a.format("efsm")     # a = std(a?{500})
0.09s  0.09s  0.09s  a.format("fado")     # a = std(a?{500})
0.09s  0.09s  0.09s  a.format("grail")    # a = std(a?{500})
0.32s  0.32s  0.31s  a.format("tikz")     # a = std(a?{500})
3.31s  3.34s  3.23s  a.infiltration(a)    # a = std([a-e]?{30})
2.67s  4.32s  2.72s  a.minimize("moore")  # a = std([a-k]{2000})
4.68s  4.43s  4.45s  a.minimize("signature") # a = std([a-g]{800})
3.97s  3.96s  3.88s  a.product(a)         # a = std([a-e]?{50})
2.35s  2.33s  2.38s  a.proper()           # a = thompson(a?{1200})
0.71s  0.70s  0.73s  a.ratexp()           # a = std([a-d]?{100})
2.26s  2.39s  2.26s  a.shortest(5)        # a = de_bruijn(9)
2.99s  2.97s  3.01s  a.shuffle(a)         # a = std([a-e]?{50})
1.26s  1.24s  1.28s  a.sort()             # a = std([a-e]?{600})
0.99s  0.98s  1.01s  a.synchronizing_word() # a = de_bruijn(6)
0.16s  0.20s  0.14s  b.format("text")     # b = lal_char(abc)_b, 100000x
0.57s  0.59s  0.57s  b.ratexp(e)          # e = "(\e+a)" * 500, 100x
3.18s  3.18s  3.16s  r.derived_term()     # r = (a+b)*b(<2>a+<2>b){150}   on [a-z]  -> Z
2.80s  3.06s  2.79s  r.derived_term()     # r = (a+b)*b(<2>a+<2>b){150}   on [ab]   -> Z
0.17s  0.18s  0.16s  r.format("text")     # r = b.ratexp("(\e+a)" * 500), 1000x
0.42s  0.43s  0.43s  r.linear()           # r = (a+b)*b(<2>a+<2>b){150}   on [a-z]  -> Z
0.04s  0.04s  0.04s  r.standard()         # r = (a+b)*b(<2>a+<2>b){20000} on [a-z]  -> Z
0.04s  0.04s  0.04s  r.thompson()         # r = (a+b)*b(<2>a+<2>b){20000} on [a-z]? -> Z
1.06s  1.11s  1.08s  read(s)              # s = dot(std(a?{500}))
0.23s  0.24s  0.25s  read(s)              # s = efsm(std(a?{500}))
0.15s  0.16s  0.16s  read(s)              # s = fado(std(a?{500}))

* vcsn/algos/product.hh (origins_): Move last.