add new subset-automaton decorator, and use it for minimization
Build subset-automaton, a decorator related to the *shape* of an
automaton rather than to the algorithm which produced it, which
reflects a slight change of directions in decorators.  Use it in
quotient to build a decorated automaton from classes -- without making
everything inefficient: the algorithm remains identical; it could have
been slightly simplified, but at the cost of introducing dynamic
lookups for state sets.

The decorator is intended to be useful for determinization as well.

Don't accept "brzozowski" as an algorithm variant at the static level;
rather, let the user call minimize_brzozowski directly (Rationale: the
decorator would be very different, and typically uninteresting).
Accept "brzozowski" as a variant at the Dyn and Python levels.

No change in score from the previous commit.  On minsky:
1.64s  1.64s  a.minimize("moore")  # a = std([a-k]{2000})
2.34s  2.32s  a.minimize("signature") # a = std([a-g]{800})

* vcsn/core/subset-automaton.hh: New.

* vcsn/algos/minimize.hh: Disable brzozowski; don't strip input.
* vcsn/algos/quotient.hh: Remove the VCSN_ORIGINS kludge, now obsolete here.

* vcsn/algos/minimize-moore.hh, vcsn/algos/minimize-signature.hh,
* vcsn/algos/minimize-weighted.hh: Change result types.

* lib/vcsn/dyn/context-parser.cc: Here.
* lib/vcsn/dyn/context-printer.cc: Here.

* vcsn/local.mk: Here.

* tests/python/minimize.py, tests/python/minimize.dir/intricate.exp.gv,
* tests/python/minimize.dir/nonlal.exp.gv,
* tests/python/minimize.dir/redundant.exp.gv,
* tests/python/minimize.dir/small-nfa.exp.gv,
* tests/python/minimize.dir/small-weighted.exp.gv,
* tests/python/local.mk: Update testsuite.

* vcsn/algos/minimize-brzozowski.hh: Write "brzozowski" with a 'w'.

* NEWS.txt: Brag.