doxygen: add support
Add doxygen support in the build system.  Minor hack in a header file:
add doxygen comments to namespaces (::vcsn and ::vcsn::dyn) in order
to force doxygen to generate documentation for all their members.  We
need this as the exported interface for algorithms is in naked
functions; having documentation on classes is not enough for Vaucanson.

change in document namespaces to force doxygen * to emit documentation
for func

* Doxyfile: New.
* doc/vaucanson-duck.png: Add logo, to be shown in the header.

* Makefile.am: Here.
* configure.ac: Here.

* doc/local.mk: New.
* vcsn/core/fwd.hh: Little hack, as explained above.