labelset: enable extensible alphabets
Eventually I'd like to have "lal_char_b" etc. denote an open labelset,
and "lal_char(...)_b" to denote a closed one (possibly empty).

There are a few atrocities in here, particularly wrt constness.  This
will have to be addressed at some point.

* vcsn/alphabets/setalpha.hh (open): New.
Provides a means to control whether unknown letters are rejected,
or added to the genset.

* vcsn/labelset/letterset.hh, vcsn/labelset/nullableset.hh,
* vcsn/labelset/oneset.hh, vcsn/labelset/tupleset.hh,
* vcsn/labelset/wordset.hh, vcsn/core/rat/ratexpset.hh,
* vcsn/core/rat/ratexpset.hxx (open): New.
Forwarded to the labelset.

* vcsn/algos/edit-automaton.hh, lib/vcsn/algos/edit-automaton.cc
(letters_): Remove.
(open): New.
Instead of trying to build the set of needed letters, open the
labelset.
* lib/vcsn/algos/efsm.cc, lib/vcsn/algos/fado.cc: We need open
labelsets.