efsm: support reading transducers
* vcsn/algos/edit-automaton.hh, lib/vcsn/algos/edit-automaton.cc: Don't
try to support any kind of tuple of labelsets: support only two-tape
automata with unknown context type.
When the context is known (which is the case of dot I/O) there is
no such limitation.

* lib/vcsn/algos/efsm.cc: When reading a transducer, announce two-tape
labels to the automaton editor.

* bin/efstdecompile: Be more robust to Open FST considering that "(a,x)*"
is an acceptor, since it looks at symbol _numbers_ only and sees "(1,1)*".

* tests/python/efsm.dir/a2xyz.efsm: New test case for "real" transducers.