coverage: workaround virtual destructors issues
lcov fails with messages such as:

[09:16:53][Step 5/6] Reading data file vcsn.info
[09:16:53][Step 5/6] Found 432 entries.
[09:16:54][Step 5/6] genhtml: ERROR: Demangled function name _ZN4vcsn3rat5unaryILNS0_6type_tE8ESt5tupleIIccEEiED0Ev  maps to different lines (53 vs 150)
[09:16:54][Step 5/6] Demangling 44720 function names
[09:16:54][Step 5/6] Makefile:4390: recipe for target 'coverage' failed
[09:16:54][Step 5/6] make: *** [coverage] Error 255
[09:16:54][Step 5/6] Process exited with code 2

where the aforementioned lines are:

    /// The abstract parameterized, root for all rational expression
    /// types.
    template <typename Label, typename Weight>
=>  class node
      : public std::enable_shared_from_this<node<Label, Weight>>
      , public exp

and

    template <exp::type_t Type, typename Label, typename Weight>
==> class unary
      : public inner<Label, Weight>

and the said symbol is:

  vcsn::rat::unary<(vcsn::rat::type_t)8, std::tuple<char, char>, int>::~unary()

Repeating the experiment reveals several other cases, all being cases
of destructors which have in common to be derived from a class with a
virtual destructor.  These destructors are needed, as these classes
are used polymorphically, but until a proper fix is found, let's
disable them in COVERAGE builds.

Based on comments by Alexandre Durat-Lutz on an earlier version of
this commit.

* vcsn/core/rat/ratexp.hh, vcsn/dyn/type-ast.hh: Disable the
virtual destructors if COVERAGE is defined.