compose: no longer offer "const" insurance on the input arguments
We have a constness issue: the dyn:: support does not take const into
account, although C++ obviously does.  As a result we might have types
equal from the dyn:: point of view corresponding to different C++
types, and therefore, we _can_ have dynamic_cast failures in dyn::
because of this.

blind_automaton is the most tricky case, as we want to use it
potentially to modify an automaton (actually, do we really?), and it's
used on a const automaton in the case of compose.  As a result we
compile code for blind_automaton wrapping a non-const automaton type,
but use code for a const wrapped automaton.

To make things simple, in a first step, let's remove the const
guarantees on the input arguments of compose.

* lib/vcsn/algos/compose.cc, python/vcsn_cxx.cc, vcsn/algos/compose.hh,
* vcsn/dyn/algos.hh (compose): No longer const on input.