dot: improve conformance with Dot
Currently we accept attributes without assignments, such as "my_node
[attribute]", which is not recognized by dot.  Besides, we don't
recognize ";" as a valid attribute separator, which it is.

Being more tolerant about Dot than dot seems ok, however, it we use
vcsn as a Dot validator, we might accept some automata input which are
rejected by dot.  This is not good if for instance we have an IPython
widget that reads a dot input, validates it by vcsn before passing the
original dot specification to dot.

* lib/vcsn/dot/parse.yy: Address these issues.
* tests/python/automaton.py: Check it.