determinize: remove complete
Benchs show that even when _not_ asking for a complete result, the
mere fact of checking whether not to complete does incur a small
performance hit.

Worse yet: the next commit completely changes to core loop of the
algorithm, and making the automaton complete really adds cruft to the
algorithm.  So instead of trying to merge two algorithms into one,
stop supporting the 'complete' feature in determinization, just call
complete() afterwards if needed.

On erebus, previous commit and then this one:

   n-1   n
  7.93s 7.80s: a.determinize()      # a = ladybird(21)
  6.64s 6.45s: a.determinize()      # a = lal(a-zA-Z0-9).ladybird(18)

* vcsn/algos/determinize.hh, lib/vcsn/algos/determinize.cc,
* python/vcsn_cxx.cc, tests/python/determinize.py
* vcsn/algos/are-equivalent.hh, bin/vcsn-tafkit.cc,
* vcsn/ctx/lal_char_b.hh, vcsn/dyn/algos.hh: Here.