determinize: introduce polynomials of... states
This is extremely bizarre at first sight, but seems to work really
well.  Our state names are nothing but weighted sets of states, which
is exactly what polynomials do.  We get a number of free services,
such as less-than comparison, and printing.  Addition and division (by
a scalar) are used in the following commit.

This shows that it would be most useful to implement Boolean
polynomials with a finite support as bitsets, and would allow to have
a single implementation of determinization for both weighted and
Boolean cases.

* vcsn/algos/determinize.hh (stateset): New.
Use it instead of an adhoc structure.