product: leave the computation of the resulting type to the caller
This is more general: the result could be put in a supertype of the
join (or the meet).  This is nicer: as of today by default we also
take the type of the join although in the case of product, we need its
meet.  This is simpler: we can remove some of the metafunctions used
to compute meet/join on tuple of automata.

* vcsn/algos/product.hh (product_automaton_impl): Take the result
automaton a parameter/argument.
Adjust dependencies.

* vcsn/algos/product.hh, vcsn/core/tuple-automaton.hh: Stop the anomaly
of not specifying all the parameters.