translate: issue a parse error when reading an invalid context
Currently, when parsing a context, we actually get ready to accept any
type (including automata, ints, or pure junk).  As a result, we also
try to compile the corresponding make_context, even for obvious parse
errors.  This is slow, and provides a bad error message.

Speed is especially an issue with the current %automaton IPython magic
line, as it tries to compile the result at each key stroke, which
results in incredibly slow interaction, as we wait for the compiler to
fail...

* vcsn/dyn/context-parser.hh, lib/vcsn/dyn/context-parser.cc
(parse_context): New.
(context_): New overload to factor some calls.
* lib/vcsn/dyn/translate.cc: Use parse_context.
Remove duplicate handling of EOF.