dot: fix the style of the named states
A state with a name, which uses shape=box and style=rounded, does not
present the same patterns than regular nodes, hence they were not
restored as rounded when colored.  It is simpler to set the style just
once as filled and rounded, with the right width, instead of trying to
fix all the different patterns.

* vcsn/algos/dot.hh: Set the patterns here for default.
* python/vcsn/dot.py: Patch them here.

Adjust the test suite (hurray for update-test).