has_twins_property: improvement
Caculate and compare the weight when visit a new state (the old weight
and new weight of the visited state from started state in a
component). Indeed, we don't use the depth first search algorithm,
just only travel all states by output transition in the component and
caculate.

Compare the time after change:
  r1 = "<3>a((<2>c<3>b<4>a)*){50}<7>c+<2>a((<3>c<2>b<4>a)*){50}<5>a"
  a1 = vcsn.context("lal_char(a-c)_q").ratexp(r1).standard()
  r2 = "<3>a((<2>c<3>b<4>a)*){50}<7>c+<2>a((<3>c<5>b<4>a)*){50}<5>a"
  a1 = vcsn.context("lal_char(a-c)_q").ratexp(r2).standard()

  n     n+1
  1.51s 1.48s
  1.36s 1.34s

* vcsn/algos/has-twins-property.hh