determinize: weighted: do not treat final states especially
Mapping post suffices, there is no need to handle the final states.
We even benefit from a small speed up.  However, this speed up does
not carry nicely to the boolean case on ladybird.  But that's because
we actually cut many many transitions: half of the resulting states is
final, and we would consider all their transitions whereas in the
current approach we see that they are final, and don't look at the
corresponding post transition.

 0.15s  0.14s: a.determinize("weighted") # a = de_bruijn(13)
 0.32s  0.30s: a.determinize("weighted") # a = de_bruijn(14)
 1.48s  1.36s: a.determinize("weighted") # a = de_bruijn(16)
 0.13s  0.13s: a.determinize("weighted") # a = std([^]{100}), c = [a-zA-Z0-9] -> B

* vcsn/algos/determinize.hh: Consider final states as regular states.