polynomialset: scalar_product
I used it to see if the weighted determinization would benefit from
considering the "polynomial" of final states, and make the scalar
product with the current polynomial.  This is how the Boolean case is
implemented, and in the case of ladybird, it is (slightly) more
efficient than looking at all_out and treating final states as regular
states.

But there was no improvement, and the current code is much better as
it is.

* vcsn/weightset/polynomialset.hh (scalar_product): New.