linear: merge into derived-term
Well, "linear" is just another implementation of the derived-term
automaton, so rather than defining another function, fuse them into a
single implementation, with "algo" selection.  While at it, after all,
breaking or non-breaking is also an choice of algorithm, so fuse this
argument into the "algo" argument.

For consistency with other algorithms, in addition to the four base
cases, support "auto", which is the default.

* tests/python/first-order.py (linearer): Fuse into....
* vcsn/algos/derived-term.hh (derived_termer): this.
(derived_term): Rename as...
(derived_term_derivation): this.
(linear): Rename as...
(derived_term_expansion): this.
* bin/vcsn-tafkit.cc, lib/vcsn/algos/derivation.cc, python/vcsn_cxx.cc
* tests/python/derivation.py, vcsn/algos/first-order.hh,
* vcsn/ctx/instantiate.hh, vcsn/dyn/algos.hh: Adjust.