has_twins_property: cycle-unambiguous
The initial version for checking a automate is cycle-unambiguous or
non. The idea based on Mr Alexandre Duret-Lutz's idea: using product
two automata to check. The algorithm is: First, it unset all initials
and finals states. Second, assigne each state to initial and final and
use the function is_ambiguous to verify. If it is ambiguous, it has at
least two diffirent cycles at this state(initial/final) with the same
input, so it is cycle-unambiguous. This is not optimal solution but it
prouves we can do it.

* vcsn/algos/has-twins-property.hh: add function is_cycle_unambiguous.
* tests/python/has-twins-property.py, python/vcsn_cxx.cc,
* lib/vcsn/algos/has-twins-property.cc, vcsn/dyn/algos.hh,
* lib/vcsn/dyn/context-printer.cc: Here.