join: remove default support for idempotence
The previous commit introduced a new implementation of join (on top of
detail::join_impl class templates) which features built-in support for
idempotence.  Remove it.

Indeed, join is idempotent on the values, but not on their types.  If
b v b is indeed b, in the case of labelsets it no longer works so
simply: letterset v letterset = letterset, but we must perform the
union of the alphabets.  RatExpSet<Ctx> v RatExpSet<Ctx> =
RatExpSet<Ctx>, but we must perform the join on the labelsets and on
the weightsets (e.g., the weights are expressions, therefore also have
an alphabet).

The current implementation is troublesome on two regards.

First, it forces us to duplicate some code.  For instance the rule for
ratexps is the same for both ratexps of different contexts, and of
same context.  However, since we already have a join_impl<T, T> that
deals with idempotence, the general case, join_impl<ratexpset<Ctx1>,
ratexpset<Ctx2>>, introduces a C++ ambiguity when Ctx1 == Ctx2, as
both specializations are equally good according the standard.  So in
order to disambiguate, we used to explicitly forbid the case Ctx1 ==
Ctx2, which in turn requires that we provide a specialization for
join_impl<ratexpset<Ctx>, ratexpset<Ctx>> (same context types).  There
comes our duplication.  And of course, this is no longer ambiguous as
this is a perfect match.

The second issue, even worse, is that the default join_impl<T, T>,
which does nothing but returning one of the input arguments, _hides_
the missing specializations.  So for instance, unless we explicitly
support letterset v letterset, it will compile, but will not perform
the union of the alphabets.

So on second thought, I prefer having to explicitly state the trivial
b v b = b for all the weightsets, but avoid code duplication, and stop
hiding missing cases.

* vcsn/core/join.hh: No longer support a default case for idempotence.
* vcsn/core/rat/ratexpset.hh: De-specialize the handling of ratexpset
vs. ratexpset.
* vcsn/ctx/context.hh: Likewise.
* vcsn/weightset/b.hh, vcsn/weightset/f2.hh, vcsn/weightset/q.hh,
* vcsn/weightset/qmp.hh, vcsn/weightset/r.hh, vcsn/weightset/z.hh,
* vcsn/weightset/zmin.hh: State idempotence.

* vcsn/labelset/nullableset.hh, vcsn/labelset/tupleset.hh,
* vcsn/weightset/polynomialset.hh (join): Reimplement properly.