gcc: 4.9.1 is fixed wrt evaluation order in list initializers
* vcsn/labelset/tupleset.hh (conv_): While the standard guarantees
the order of evaluation for list initializers, it says nothing about
order of evaluation in perfect forwarding.  In particular, there is
no reason for make_tuple(args...) to evaluate left-to-right, but
tuple<Args...>(args...) is well defined.  And now that gcc is fixed,
we must not use the latter, and the configure.ac test does check for
it, not for the behavior of make_tuple.

In short, don't ask for more than what the standard provides.