cycle_ambiguous: fixes
Change the name of function: from is_cycle_unambiguous to
is_cycle_ambiguous (the result was already right, only the name of the
function was negated).

Use require() to assure the preconditions: automaton is not cycle
ambiguous in has_twins_property().  Change comment to describe better
the idea to find cycle ambiguous.

* vcsn/algos/has-twins-property.hh,
* tests/python/has-twins-property.py, bin/vcsn-score,
* lib/vcsn/algos/has-twins-property.cc, vcsn/dyn/algos.hh,
* lib/vcsn/dyn/context-printer.cc, python/vcsn_cxx.cc: Adjust.