determinize: support the case of Z
Originally we used to normalize how states by ensuring that the sum of
the weights is 1, i.e., the norm is the sum of the weights.  This is
not nice when the sum is 0.  We now use the first weight as norm,
i.e., the normalized state has 1 as first weight.  This is ok for
division rings, but for Z, this is inappropriate.  In this case, the
norm is taken to be the gcd of the weights, but with the sign of the
first weight.

* vcsn/weightset/polynomialset.hh (norm_, normalize_here): Do that.
* tests/python/determinize.dir/z-det.gv, tests/python/determinize.dir/z.gv,
* tests/python/determinize.py: Check that.