build: pass --no-as-needed to the linker when needed
On some distros shared libraries that appear to be useless (i.e., they
don't provide symbols that were actually used by the object that was
linked against them) are not kept as dependencies, therefore they are
not loaded, and their initialization section is not executed either.

This is bad for Vaucanson which uses the libraries' initialization to
register their content to the main executable.  In other words,
because we perform the "linking" ourselves (via our registers), the
linker does not see that we need these libraries.

So find if this compiler set up actually discards the "useless"
libraries, and then provide a means to force the loading of such
libraries.

Reported by Victor Marsault and Valentin Tolmer.

* build-aux/m4/vcsn-as-needed.m4: New.
* configure.ac: Use it.