ipython: improve %automaton edition
The format can be chosen among those supported by Vaucanson: dot,
daut, efsm, etc.

The widgets can be placed horizontally or vertically with the argument
h or v.  Horizontally the text box is on the left, the automaton on
the right and the erromessage under the text box.  Vertically the text
box is under the automaton and the error message under the text box.
This modifications allows the user to switch the visual mode when the
automaton become to long for the cell.

* python/vcsn/ipython.py: Here.