taf-kit: stop sorting the output
Calling sort is sometimes hiding some information on the resulting
automaton (by, for instance, adding state labels when there are none).
Sorting in TAF-Kit was done in order to help the user get something
which is likely to be what will be parsed if used in input, but
really, using TAF-Kit here is not right.

* bin/parse-args.cc: Here.