transpose: don't forget to transpose the number of inital/final states
In e3a77da6cfe3e1dd09f31ab921eeaf27799f542d I "fixed" a bug, which I
did not understand well, and indeed, it was not what I thought (an
issue with a temporary value being dead when used) but could not
understand (this temporary value had no reason to be dead).

Canh Luu found and reported the real bug: because we forgot to
transpose num_initials and num_finals, we could believe that there was
one final state when there is actually none, so we crash when trying
to read this imaginary final state.

I will, however, maintain the changes from the aforementioned commit,
as it is also easier to read this way.

* tests/python/transpose.py: Check.
While at it, make sure that transpose is really an idempotence, not
simply that transpose(transpose(a)) looks like a.
* vcsn/algos/transpose.hh: Fix.