transpose: fix an annoying issue: strip de-transposes but does not strip
This has long been a problem in Vaucanson:
a.determinize().transpose().strip() returns exactly a.determinize().
Yes, the result is wrong as (i) we lost the transposition and (ii) we
have kept the decoration due to determinization.  The real result
should be that of a.determinize().strip().transpose().

Reported by Canh Luu.

* vcsn/algos/transpose.hh (naked_automaton): New.
(strip): New.