standard, thompson: fix the type dependencies
* vcsn/algos/standard.hh, vcsn/algos/thompson.hh:
These algorithm were parameterized by Context, which is wrong:
we need to parameters: Aut, in order to build an automaton of
that type, and RatExpSet, in order to be able to read the ratexp
we are converting.

Previously, we were using context as a common type for both,
building a ratexpset inside the visitors.  This was wrong,
keep the valuesets at hand.