rat: let ratexp be parameterized by the Context instead of Label, Weight
The rational expressions are values, and as such, they are "stupid",
and just "contain" some value without knowing how to interpret them
(very much the same way as "bool" weights do not know whether they
below to F2 or to B).

However, I now wish I could perform hash-consing on the ratexp, which
means that I need to be able to hash and equality-compare the ratexps,
which means I need to be able to hash the labels and weights, which is
something that only labelset and weightset can do, as of today.

So I have two options:

a. have the ratexp know the labelset and weightset (their types
   suffices, we don't need a value as the hash function is static in
   valuesets), i.e., have the ratexp know its context type.

b. move the hash and equality-compare functions from valueset to
   value.  This seems wrong, as it would prevent the valueset from
   using equivalences to between values, or use a specific order.

* vcsn/core/rat/ratexp.hh: Do that.
Adjust all dependencies.