doc: provide upload-doc for notebooks
I change, again, my point of view.  I dislike that Automake (and Make
actually) be such pains in the neck when there are spaces in the file
names.  I worked around this by adding quotes to the file names, but
them neither Make nor Automake recognized them, and I had to do
everything by hand.  Now it's even worse when I try to add Make rules
to generate HTML from notebooks.

Stop this.  Don't use spaces in names, and enjoy the benefits of the
good old Automake/Make tandem.

* doc/local.mk (RSYNCFLAGS): Delete excluded.
* share/vcsn/notebooks/Editing Automata.ipynb: Rename as...
* share/vcsn/notebooks/Editing-Automata.ipynb: this.
* share/vcsn/notebooks/Some developer documentation.ipynb: Rename as...
* share/vcsn/notebooks/Hacking.ipynb: this.
* share/vcsn/notebooks/index.ipynb, share/vcsn/notebooks/local.mk: Adjust.

But now "make install" fails because our notebooks_DATA has directory,
and these directories are not kept when installed, so for install
index.ipynb and algos/index.ipynb collide.  Automake offer nobase, but
it is too coarse, and it would also keep the "share/vcsn" part of the
source directory.

Well, KISS: drop the hierarchy.

* share/vcsn/notebooks/algos/*: Move to...
* share/vcsn/notebooks: here.

But now that we use nbconvert, we have to make sure it works...

* configure.ac (NBCONVERT, HAVE_NBCONVERT): New.
* share/vcsn/notebooks/local.mk: Use them.