score: update contexts style
This shows a serious regression in the performances of simply printing
a context name (context.format('text') in Python).  I had not realized
that during the overhaul of the syntax of the contexts.

  0.16  0.16  0.16  0.25  b.format("text") # b = [abc] -> B, 100000x

There is no real reason for this printing to be slower, as the
algorithm has not changed, and the result is the same.  What has
changed, though, is the cost of the computation of the signature used
to perform the dyn:: dispatch.

So in clear, this regression is a regression of the dyn::
polymorphism: it is quite sensitive to the syntax of the string used
as a key.  We used to hash things like

    lal_char(abc)_b

and now it is

    letterset<char_letters(abc)>, b

i.e., from 15 chars to 31.

* bin/vcsn-score: Use [abc] -> B as displayed syntax.
* bin/vcsn-score-compare: Adjust the past.