labelset: revamp the parsing routines
Currently there is too much code duplication between the conv
routines.  One problem, for instance, is that \-escapes are handled in
different places, instead of a single place.  This is due to the fact
that we don't make the difference, so far, between getting a char, and
getting a letter which is a char (which might need two chars: \\
denotes a single backslash).

In order to be more readable, use a more rational naming scheme:
get_char, get_letter, and get_word.

Also, when reading a word, be eager, instead of lazy: instead of
"peeking" at the rest of the stream, catch all the letters that come,
and complain afterwards if some were not in the alphabet.  The main
motivation here is UTF-8: "peeking" no longer works, as we may need a
large lookahead (some letters are multibyte).  So we have to really
get a letter (possibly multibyte), and cannot use putback either, as
it is guaranteed for a single byte only.

This simplifies the case of reading polynomials which used to perform
try-catch on domain_error, which was sometimes really a PITA, as some
errors were caught, unexpectedly for the developer.

* vcsn/alphabets/char.hh (get_char, get_letter): New.
* vcsn/alphabets/setalpha.hh (make): Use get_letter, instead of duplicating the
handling of escapes or other idiosyncrasies from the letters.
(conv): Rename as...
(get_word): this to be clearer.
Also, be eager (see comment above).
Use raise/require instead of domain_error.
* vcsn/labelset/genset-labelset.hh (convs_): Use get_letter instead of
duplicating, again.
* vcsn/labelset/letterset.hh, vcsn/labelset/wordset.hh (conv): Adjust: use get_letter
or get_word.
* vcsn/weightset/polynomialset.hh: Likewise.