alphabet: prepare to support letter types that are not iterable
We would like to support "alphabets" that are composed of "words".
This is not "wordset", but really another type of "letters": strings.
However strings do not support a natural iteration concept.  So our
support for intervals (e.g., lal_char(a-z)) does not work properly on
strings (what is "lal_string('aa'-'zzz')?").

So be ready to not support ranges.

* vcsn/alphabets/setalpha.hh (has_range, add_range): New.
(make): Use it.