context: remove a useless constructor
This constructor:

    /// Build a context whose labelset constructor takes no argument.
    template <typename LabelSet2 = labelset_t>
    context()
      : context{typename std::enable_if<is_lao, labelset_t>::type{},
                weightset_t{}}
    {}

is wrong: (i) the enable_if is not about a template parameter, so it
does not disable anything, and besides, (ii) the enable_if is not in
the signature part, so it does not disable anything at all.  Actually
we were only saved by overloading, and lazy instantiation of
templates.

It should have looked like this:

    /// Build a context whose labelset constructor takes no argument.
    template <typename LabelSet2 = labelset_t>
    context(if_lao_t<LabelSet2>* = 0)
      : context(labelset_t{}, weightset_t{})
    {}

However, it does not appear to save us from anything to accept default
construction of lao only.  For instance, the empty lal_char is useful
(but weird indeed).

* vcsn/ctx/context.hh (context): Simplify.
* vcsn/core/kind.hh (if_la*): Remove, now useless.