build: when using hidden visibility, use intermodule_holder for flyweights
Interestingly demo/prod-eval is currently the only place where we
check two 'symbol's from different dl-modules.  This requires that we
use a single symbol holder for the whole program, not just one per
shared library.

This fails when we use ELF visibility control (the linker no longer
fuses the different static_holders), hence we must use use
Boost.Flyweight's intermodule_holder.  But since this is costly
(according to the documentation), we do it only when required.

Unfortunately this pulls in some headers such as ucred.h on OS X,
which features a "struct label;".  So we have to qualify some of our
"label"s.

* configure.ac (HAVE_VISIBILITY_HIDDEN): Check it.
* vcsn/misc/symbol.hh: Use it to select intermodule_holder.
(operator+): Make it easier to concatenate strings and symbols.
Will be used often to build snames.
* tests/demo/local.mk: Fix dependency.
* tests/demo/prod-eval.cc: Add comments.