mutable_automaton: pass labels by value
Our typical labels are char or maybe Flyweights, so passing them by
const-ref is not smart.  Worse yet: in the case of the
transpose_automaton, we sometimes pass transposed-labels, i.e.,
temporary values.  Some of our functions return lambdas that capture
labels, and therefore we have the serious risk of depending on dead
temporaries.

As a matter of fact, this commit was made because of the following one
that made this error.  Arguably, it is the lambda that should capture
by copy, the problem is not that the label_t was passed by cref.
However, for consistency, let's stick to label_t everywhere.

If that's a problem with law_char, then we should probably move to
using Flyweights there too.

* vcsn/algos/filter.hh, vcsn/algos/focus.hh, vcsn/core/expression-automaton.hh,
* vcsn/core/mutable-automaton.hh: Pass labels by copy.