dyn: reduce the code duplication to strip gensets
When given a context specification (e.g., "lal_char(ab), b") we
sometimes need the corresponding symbol without generators ("lal_char,
b").  This is achieved by dyn::context_base::sname (which is a weird
location for such a feature), but it is also acheived by the
signature-parser/signature-printer.

Use only the latter.

* vcsn/dyn/signature-printer.hh (normalize): Be ready to normalize
with or without runtime values.
* lib/vcsn/dyn/signature-printer.cc: Do it.
* lib/vcsn/algos/make-context.cc (make_context): Use normalize instead
of sname().
* lib/vcsn/dyn/context-parser.cc (genset_): Support backslashes.
* lib/vcsn/dyn/translate.cc: Don't use sname, actually this was unneeded.
* vcsn/dyn/context.hh (sname): Remove.
* lib/vcsn/dyn/context.cc: Remove.