tests: filter: make sure "eval" works
At some point, it did not work.
Reported by Pierre Senellart.

The test did reveal a problem though: detail::max_forward requires a
non empty range, which is not the case with states().  So let's use
all_states().  Besides, we now that the states are returned in order,
so we just need the last element, we don't need to compare them.

* vcsn/misc/algorithm.hh (back): New.
* vcsn/algos/eval.hh: Use it.
* tests/python/filter.py: Check filter_automaton.eval.