read: fix the order of the arguments
The print routines have all been normalized to print(ValueSet, Value,
Stream), or print(Value, Stream) for the dynamic case, the logic being
that after all, neither the ValueSet not Value can be left
unspecified, while the Stream can default to std::cout.

In addition, this is also more consistent, as the ValueSet is
otherwise always the initial argument.

Let's apply the same change to the read_* functions.