entries: fix an amazing bug in rendering entries with \e
I am amazed that we never saw that bug before.  It seems to mean that
it is rare to have a spontaneous and a non-spontaneous transition
between two states...

Still, when we read "\e, a, b, c" as an entry, we display "\e, [\ea-c]"!!!

While working on this, I also changed the "same weight rule":
previously "\e, <2>[a-d]" was displayed "\e, <2>a, <2>b, <2>c, <2>d".
Now it displays "\e, <2>[a-d]".  However, if the letters have
different weights, then we don't use any range.

* vcsn/weightset/polynomialset.hh (print_with_ranges_): Address
these issues.
* tests/python/automaton.py: Check that.