nullableset: fix broken implementation
Currently nullableset<LabelSet> derives from LabelSet!  It makes no
sense.  For instance, it means that LAN is a good implementation of
LAL...

Because of this, several bugs remained hidden.  For instance, after a
call to proper_here, how come the conversion LAN to LAL work although
there is no conv support for this?  Because the LAN was considered a
valid LAL...

* vcsn/labelset/nullableset.hh (nullable_helper): It does not need
to know the nullableset which uses it (null).
This makes the code much easier to understand, as there are less
mutual dependencies.
Remove useless code.
(nullableset): Fix many issues.
In particular, do _not_ derive from the LabelSet we wrap.
(nullableset::genset, nullableset::letters_of): New.
Fix issues wrt handling "one" in routines such as transposition.
* vcsn/labelset/letterset.hh (conv): Add support for LAN to LAL.